// src/app/models/course-management/courseInstructorData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CourseInstructorData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** custom รูปคน */
    base64?: string;
    /** FK (course_main) */
    courseId?: number;
    /** PK */
    courseInstructorId?: number;
    /** custom อีเมล */
    email?: string;
    /** อีเมล */
    externalEmail?: string;
    /** ชื่ออาจารย์/ผู้รับผิดชอบ  (ภาษาอังกฤษ) */
    externalNameEn?: string;
    /** ชื่ออาจารย์/ผู้รับผิดชอบ (ภาษาไทย) */
    externalNameTh?: string;
    /** FileName */
    filename?: string;
    first?: number;
    /** custom ชื่ออาจารย์  (ภาษาอังกฤษ) */
    fullnameEn?: string;
    /** custom ชื่ออาจารย์ (ภาษาไทย) */
    fullnameTh?: string;
    /** FK (mas_personal)  */
    instructorId?: number;
    /** อาจารย์/ผู้รับผิดชอบหลักสูตร (มีผู้รับผิดชอบหลักได้คนเดียวต่อ 1 หลักสูตร)  */
    instructorMain: boolean;
    /** บุคลากรภายนอก */
    instructorType?: boolean;
    mode?: string;
    /** Module */
    module?: number;
    /** Prefix */
    prefix?: string;
    rowNum?: number;
    size?: number;
}